Actualizo

+++ src/UART/UART.h
@@ -4,7 +4,7 @@
 #define UART_UART_H_
 
 void UART4_Config(General *Gen);
-void usart2SendData(uint8_t *Buff, uint8_t Length);
+void usart2SendData(uint16_t *Buff, uint8_t Length);
 void readMessage(General *Gen);
 void Init_MenuUART(General *Gen);
 void UART4_IRQHandler(General *Gen);
